rename utils (#176)
* refactor: cli_utils -> cli

* refactor: cli_utils -> cli

* refactor: rule_utils -> rule

* refactor: rule_utils -> rule
