+++ BALSAMIC/snakemake_rules/align/samtools.rule
@@ -3,7 +3,7 @@
 
 __author__ = "Hassan Foroughi Asl"
 
-from BALSAMIC.utils.rule_utils import get_conda_env
+from BALSAMIC.utils.rule import get_conda_env
 
 
 rule samtools_sort_index:
