+++ BALSAMIC/snakemake_rules/quality_control/collectqc_single.rule
@@ -3,7 +3,7 @@
 
 __author__ = "Hassan Foroughi Asl"
 
-from BALSAMIC.utils.rule_utils import get_conda_env, get_picard_mrkdup
+from BALSAMIC.utils.rule import get_conda_env, get_picard_mrkdup
 
 picarddup = get_picard_mrkdup(config)
