+++ BALSAMIC/snakemake_rules/quality_control/cutadapt.rule
@@ -4,7 +4,7 @@
 
 __author__ = "Hassan Foroughi Asl"
 
-from BALSAMIC.utils.rule_utils import get_conda_env
+from BALSAMIC.utils.rule import get_conda_env
 
 
 # remove adapter from paired end reads
