+++ BALSAMIC/snakemake_rules/variant_calling/bcftools.rule
@@ -1,11 +1,11 @@
 # vim: syntax=python tabstop=4 expandtab
 # coding: utf-8
 
-from BALSAMIC.utils.rule_utils import get_vcf_files
+from BALSAMIC.utils.rule import get_vcf_files
 
 __author__ = "Hassan Foroughi Asl"
 
-from BALSAMIC.utils.rule_utils import get_conda_env
+from BALSAMIC.utils.rule import get_conda_env
 
 
 var_type = "SNV"
