+++ BALSAMIC/snakemake_rules/variant_calling/vardict_single.rule
@@ -3,7 +3,7 @@
 
 __author__ = "Hassan Foroughi Asl"
 
-from BALSAMIC.utils.rule_utils import get_conda_env, get_chrom
+from BALSAMIC.utils.rule import get_conda_env, get_chrom
 
 
 var_caller = "vardict"
