+++ tests/commands/config/test_config_sample.py
@@ -5,7 +5,7 @@ import pytest
 from BALSAMIC.commands.config.sample import get_config, write_json, merge_json, \
     set_panel_bed, get_output_config, get_sample_config, get_analysis_type, \
     check_exist
-from BALSAMIC.utils.cli_utils import iterdict
+from BALSAMIC.utils.cli import iterdict
 
 
 def test_get_config(config_files):
